requirements.txt
kafka-python
praw
langdetect
iso-639
nltk
multi_rake
python-dotenv
psycopg2
matplotlib

